refactor(user): split up sub routes
